pyproject.toml
[tool.poetry]
name = "python-ha-project"
version = "0.1.0"
description = ""
authors = ["chawanghyeon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.0"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pymongo = "^4.7.1"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
django-query-counter = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
mypy = "^1.4.1"
pre-commit = "^3.3.2"
black = "^23.3.0"
autoflake = "^2.1.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
autopep8 = "^2.0.2"
bandit = "^1.7.5"
pyupgrade = "^3.7.0"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.bandit]
exclude_dirs = [".env", "tests", "**/tests/**"]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
allow_untyped_decorators = true
pretty = true
show_error_codes = true
implicit_reexport = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[[tool.mypy.overrides]]
module = [
    "tests/*",
]
ignore_errors = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
